Fixed Settings and Added a 'Reset' Button.
Settings was broken in that the app only updated function states every
time a new Settings object was created (every time the settings menu was
re-entered).  Added getFunctionStates() to allow the app to update the
state of each function (active or inactive) every time any of the
function checkboxes are checked or unchecked. Also added a 'Reset'
button, which can be pressed to return the functions to their default
state, active.