Made Timer Functional
The timer works now and shows the user how much time remains in their
quiz in the response window. However the timer pauses when the user
exits to the list of questions, a problem I'll fix eventually. Also
added the interface to make the quiz duration adjustable in the
Settings. That will be implemented in the future.