Fixed Timer
The timer now both runs and doesn't crash while the user is viewing the
question list. I made a button to move from one question to the next but
I don't have time right now to implement it. So the "Next" button is in
the Response Window, but doesn't work yet.