Fixed Timer Length
Updated the timer length to display exactly 3:00 at the start of a quiz.
Reverted my last commit because I checked again and realized the Timer
still had problems that I hadn't fixed. Commented out the Toast text and
the Looper.prepareMainLooper() for the time being so that the app can
work. They can be uncommented and fixed later.