Completed the pilgrimage to the holy gradle
I moved around a bunch of files and added some gradle scripts so that it would be able to build on gradle. I created a new application module, speedtrig, which is the main module for the app. Gradle version is actually 2.3 (woops, 2.3.1 doesn't actually exist). Gradle wrappers were also added to the root directory of the project.