Finished Navigation Drawer
Finished the Navigation Drawer. Currently there are five sections of the
app that can be navigated to using the Navigation Drawer: "Quizzes",
"Settings", "Help", "About" and "Credits". Also added some icons to
represent each of these sections.  From this point forward any future
sections of the app will be navigated to using this Navigation Drawer.