Minor Changes
On advice from Mr. Stein, added a "#" before the question number on
every quiz number to make it clear what that number is. Also changed the
Settings and About icons (the rest will also be changed soon) and
limited the quiz duration SeekBar to incrementing in values of 30
because quiz durations more precise than that seem unnecessary.