Added New Music, Made General Settings Work
The options in the "General Settings" section of the Settings menu now
function correctly. Also replaced the music with something more upbeat.
In my previous commit I said I was working on a way to optimize the way
the background music is played but I think the current system is good
enough for the time being.